+++ fmi-export-derive/src/tests.rs
@@ -1,4 +1,4 @@
-use crate::model::Model;
+use crate::{model::Model, model_variables::build_model_variables};
 use fmi::fmi3::schema;
 use schema::AbstractVariableTrait;
 use syn::parse_quote;
@@ -77,91 +77,77 @@ fn test_comprehensive_datatype_support() {
     };
 
     let model = Model::from(input);
-    let fmi_description = schema::Fmi3ModelDescription::try_from(model).unwrap();
-
-    // Test model-level attributes
-    assert_eq!(fmi_description.model_name, "ComprehensiveModel");
-    assert_eq!(
-        fmi_description.description,
-        Some("A comprehensive FMI model demonstrating all supported datatypes".to_string())
-    );
+    let model_variables =
+        build_model_variables(&model.fields).expect("Failed to build model variables");
 
     // Test variable counts - each alias creates an additional variable
-    assert_eq!(fmi_description.model_variables.float32.len(), 1);
-    assert_eq!(fmi_description.model_variables.float64.len(), 3); // velocity_f64 + velocity_alias + vel_alias
-    assert_eq!(fmi_description.model_variables.int8.len(), 1);
-    assert_eq!(fmi_description.model_variables.int16.len(), 1);
-    assert_eq!(fmi_description.model_variables.int32.len(), 3); // count_i32 + counter_alias + count_alias
-    assert_eq!(fmi_description.model_variables.int64.len(), 1);
-    assert_eq!(fmi_description.model_variables.uint8.len(), 1);
-    assert_eq!(fmi_description.model_variables.uint16.len(), 1);
-    assert_eq!(fmi_description.model_variables.uint32.len(), 1);
-    assert_eq!(fmi_description.model_variables.uint64.len(), 1);
-    assert_eq!(fmi_description.model_variables.boolean.len(), 1);
-    assert_eq!(fmi_description.model_variables.string.len(), 1);
+    assert_eq!(model_variables.float32.len(), 1);
+    assert_eq!(model_variables.float64.len(), 3); // velocity_f64 + velocity_alias + vel_alias
+    assert_eq!(model_variables.int8.len(), 1);
+    assert_eq!(model_variables.int16.len(), 1);
+    assert_eq!(model_variables.int32.len(), 3); // count_i32 + counter_alias + count_alias
+    assert_eq!(model_variables.int64.len(), 1);
+    assert_eq!(model_variables.uint8.len(), 1);
+    assert_eq!(model_variables.uint16.len(), 1);
+    assert_eq!(model_variables.uint32.len(), 1);
+    assert_eq!(model_variables.uint64.len(), 1);
+    assert_eq!(model_variables.boolean.len(), 1);
+    assert_eq!(model_variables.string.len(), 1);
 
     // Total should be 16 variables (12 base + 2 aliases + 2 additional variables with aliases = 16)
-    assert_eq!(fmi_description.model_variables.len(), 16);
+    assert_eq!(model_variables.len(), 16);
 
     // Test specific variable properties
     // Float types use Vec<T> for start values
-    assert_eq!(fmi_description.model_variables.float32[0].start, vec![1.5]);
-    assert_eq!(fmi_description.model_variables.float64[0].start, vec![2.7]);
+    assert_eq!(model_variables.float32[0].start, vec![1.5]);
+    assert_eq!(model_variables.float64[0].start, vec![2.7]);
 
     // Integer types use Option<T> for start values
-    assert_eq!(fmi_description.model_variables.int8[0].start, Some(10));
-    assert_eq!(fmi_description.model_variables.int16[0].start, Some(1000));
-    assert_eq!(fmi_description.model_variables.int32[0].start, Some(50000));
-    assert_eq!(
-        fmi_description.model_variables.int64[0].start,
-        Some(9000000000)
-    );
-    assert_eq!(fmi_description.model_variables.uint8[0].start, Some(255));
-    assert_eq!(fmi_description.model_variables.uint16[0].start, Some(8080));
-    assert_eq!(fmi_description.model_variables.uint32[0].start, Some(1024));
-    assert_eq!(
-        fmi_description.model_variables.uint64[0].start,
-        Some(1234567890123)
-    );
+    assert_eq!(model_variables.int8[0].start, Some(10));
+    assert_eq!(model_variables.int16[0].start, Some(1000));
+    assert_eq!(model_variables.int32[0].start, Some(50000));
+    assert_eq!(model_variables.int64[0].start, Some(9000000000));
+    assert_eq!(model_variables.uint8[0].start, Some(255));
+    assert_eq!(model_variables.uint16[0].start, Some(8080));
+    assert_eq!(model_variables.uint32[0].start, Some(1024));
+    assert_eq!(model_variables.uint64[0].start, Some(1234567890123));
 
     // Boolean uses Vec<bool>
-    assert_eq!(fmi_description.model_variables.boolean[0].start, vec![true]);
+    assert_eq!(model_variables.boolean[0].start, vec![true]);
 
     // String uses Vec<StringStart>
-    let string_values: Vec<&str> = fmi_description.model_variables.string[0].start().collect();
+    let string_values: Vec<&str> = model_variables.string[0].start().collect();
     assert_eq!(string_values, vec!["ComprehensiveModel"]);
 
     // Test causalities are preserved
     assert_eq!(
-        fmi_description.model_variables.float32[0].causality(),
+        model_variables.float32[0].causality(),
         schema::Causality::Output
     );
     assert_eq!(
-        fmi_description.model_variables.int32[0].causality(),
+        model_variables.int32[0].causality(),
         schema::Causality::Parameter
     );
     assert_eq!(
-        fmi_description.model_variables.uint8[0].causality(),
+        model_variables.uint8[0].causality(),
         schema::Causality::Input
     );
     assert_eq!(
-        fmi_description.model_variables.boolean[0].causality(),
+        model_variables.boolean[0].causality(),
         schema::Causality::Input
     );
 
     // Test variability assignments
     assert_eq!(
-        fmi_description.model_variables.float32[0].variability(),
+        model_variables.float32[0].variability(),
         schema::Variability::Continuous
     );
     assert_eq!(
-        fmi_description.model_variables.int32[0].variability(),
+        model_variables.int32[0].variability(),
         schema::Variability::Discrete
     );
     assert_eq!(
-        fmi_description.model_variables.boolean[0].variability(),
+        model_variables.boolean[0].variability(),
         schema::Variability::Discrete
     );
-
-    println!("✅ All comprehensive datatype tests passed!");
 }
