+++ fmi-schema/Cargo.toml
@@ -21,6 +21,8 @@ fmi2 = []
 fmi3 = []
 ## Enable support for Apache Arrow Schema
 arrow = ["dep:arrow"]
+## Enable serde support for deserialization
+serde = ["dep:serde"]
 
 [dependencies]
 arrow = { workspace = true, optional = true }
@@ -29,6 +31,7 @@ document-features = { workspace = true }
 itertools = "0.14"
 lenient_semver = "0.4"
 semver = "1.0"
+serde = { version = "1.0", features = ["derive"], optional = true }
 thiserror = { workspace = true }
 yaserde = "0.12"
 yaserde_derive = "0.12"
